requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
PyMySQL==1.1.0
DBUtils==3.0.3 